+++ package.json
@@ -41,7 +41,6 @@
     "gulp-changed": "3.2.0",
     "gulp-clean-css": "3.9.4",
     "gulp-concat": "2.6.1",
-    "gulp-eslint": "4.0.1",
     "gulp-header": "2.0.5",
     "gulp-imagemin": "4.1.0",
     "gulp-less": "4.0.0",
@@ -53,16 +52,10 @@
     "gulp-replace": "0.6.1",
     "gulp-sourcemaps": "2.6.4",
     "gulp-uglify-es": "1.0.1",
-    "is-reachable": "2.4.0",
     "lightbox2": "2.10.0",
-    "localtunnel": "1.9.0",
     "lodash.throttle": "4.1.1",
     "postcss-unmq": "1.0.2",
-    "psi": "3.1.0",
-    "require-dir": "1.0.0",
     "sauce-connect-tunnel": "0.2.1",
-    "svg-inline-loader": "0.8.0",
-    "through2": "2.0.3",
     "timelinejs": "2.35.6",
     "turbolinks": "5.1.1",
     "uglifyjs-webpack-plugin": "1.2.5",
@@ -77,11 +70,14 @@
     "chai-as-promised": "7.1.1",
     "cucumber": "4.2.1",
     "fancy-log": "1.3.2",
+    "gulp-eslint": "4.0.1",
+    "is-reachable": "2.4.0",
     "jasmine-reporters": "2.3.1",
     "jasmine-spec-reporter": "4.2.1",
     "jest": "22.4.3",
     "jest-cli": "22.4.3",
     "jsdoc": "3.5.5",
+    "localtunnel": "1.9.0",
     "merge-stream": "1.0.1",
     "minimist": "1.2.0",
     "mkdirp": "0.5.1",
@@ -89,8 +85,11 @@
     "protractor-accessibility-plugin": "github:cfpb/protractor-accessibility-plugin#on-page-load",
     "protractor-cucumber-framework": "5.0.0",
     "protractor-retry": "1.2.0",
+    "psi": "3.1.0",
     "selenium-webdriver": "4.0.0-alpha.1",
     "snyk": "1.80.1",
+    "svg-inline-loader": "0.8.0",
+    "through2": "2.0.3",
     "wcag": "0.3.0",
     "yargs": "11.0.0"
   },
